v2
tambah tabel poli

+++ app/Http/Controllers/DokterController.php
@@ -101,6 +101,10 @@ class DokterController extends Controller
     public function destroy(string $id)
     {
         $dokter = \App\Models\Dokter::findOrFail($id);
+        if ($dokter->administrasi->count() >= 1) {
+            flash('Data tidak bisa dihapus karena sudah digunakan');
+            return back();
+        }
         $dokter->delete();
         flash('Data berhasil dihapus');
         return back();
